requirements.txt
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
Deprecated==1.2.18
fonttools==4.60.1
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
networkx==2.8.8
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pip==24.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
tabulate==0.8.10
threadpoolctl==3.6.0
tqdm==4.67.1
tsplib95==0.7.1
wrapt==1.17.3
google-generativeai==0.8.3
python-dotenv==1.0.1
